+++ src/plugin/WebAudioPlugin/WebAudioPlugin.ts
@@ -46,9 +46,10 @@ export class WebAudioPlugin implements AudioPlugin {
 		system: pdi.AudioSystem,
 		loop: boolean,
 		hint: pdi.AudioAssetHint,
-		offset: number
+		offset: number,
+		loopOffset: number | undefined,
 	): AudioAsset {
-		return new WebAudioAsset(id, assetPath, duration, system, loop, hint, offset);
+		return new WebAudioAsset(id, assetPath, duration, system, loop, hint, offset, loopOffset);
 	}
 
 	createPlayer(system: pdi.AudioSystem, manager: AudioManager): AudioPlayer {